+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ### Improvements
 - Configure logging to log unhandled exceptions to console even when DEBUG=False (#835)
+- Make hc.lib.emails raise exceptions when EMAIL_ settings are not set
 
 ## v2.9.2 - 2023-06-05